+++ lib/src/common/component/calendar_picker_view_widget.dart
@@ -126,6 +126,7 @@ class _CalendarPickerViewWidgetState extends State<CalendarPickerViewWidget> {
           days: _days,
           onChanged: widget.onChanged,
           initDate: widget.initDate,
+          pickType: widget.pickType,
         ),
       ],
     );